Force local APIC physical-destination mode in the ACPI FADT table.
This matches the setting in the flags field of the bhyve-created
FADT, and also fixes an issue with the Linux 4.15 kernel when, unless
this is set, the IOAPIC gets reprogrammed to ID 0 and vCPUs starting
at 1, resulting in MSI interrupts being mis-delivered and i/o devices
timing out.

Tested with Linux distros with >= 4.15 kernels (Ubuntu 18.04, Arch
2018.03.01) and no regression with install/boot tests of Win 7/8.1/10
and server 2k16.

(issue originally reported in the FreeNAS forums)